+++ python/tests/test_classify.py
@@ -4,14 +4,13 @@ import polars as pl
 import shutil
 import os
 
-if os._exists("_test_cache"): shutil.rmtree("_test_cache", ignore_errors=True)
-
 
 def test_classify():
     df = pl.DataFrame(dict(x=["one", "two", "three"]))
     df.llm.use("test", "echo", _cache="_test_cache")
     x = df.llm.classify("x", ["one", "two"])
     assert pull(x, "classify") == ["one", "two", None]
+    shutil.rmtree("_test_cache", ignore_errors=True)
 
 
 def test_classify_dict():
@@ -19,6 +18,7 @@ def test_classify_dict():
     df.llm.use("test", "echo", _cache="_test_cache")
     x = df.llm.classify("x", {"one": 1, "two": 2})
     assert pull(x, "classify") == [1, 2, None]
+    shutil.rmtree("_test_cache", ignore_errors=True)
 
 
 def pull(df, col):