+++ python/tests/test_summarize.py
@@ -4,8 +4,6 @@ import polars as pl
 import shutil
 import os
 
-if os._exists("_test_cache"): shutil.rmtree("_test_cache", ignore_errors=True)
-
 
 def test_summarize_prompt():
     df = pl.DataFrame(dict(x="x"))
@@ -15,6 +13,7 @@ def test_summarize_prompt():
         x["summary"][0]
         == "You are a helpful summarization engine. Your answer will contain no no capitalization and no explanations. Return no more than 10 words.   The answer is the summary of the following text:\n{}"
     )
+    shutil.rmtree("_test_cache", ignore_errors=True)
 
 
 def test_summarize_max():
@@ -25,3 +24,4 @@ def test_summarize_max():
         x["summary"][0]
         == "You are a helpful summarization engine. Your answer will contain no no capitalization and no explanations. Return no more than 5 words.   The answer is the summary of the following text:\n{}"
     )
+    shutil.rmtree("_test_cache", ignore_errors=True)