+++ python/tests/test_translate.py
@@ -4,8 +4,6 @@ import polars as pl
 import shutil
 import os
 
-if os._exists("_test_cache"): shutil.rmtree("_test_cache", ignore_errors=True)
-
 
 def test_translate_prompt():
     df = pl.DataFrame(dict(x="x"))
@@ -15,3 +13,4 @@ def test_translate_prompt():
         x["translation"][0]
         == "You are a helpful translation engine. You will return only the translation text, no explanations. The target language to translate to is: spanish.   The answer is the translation of the following text:\n{}"
     )
+    shutil.rmtree("_test_cache", ignore_errors=True)