+++ r/tests/testthat/test-llm-use.R
@@ -21,7 +21,7 @@ test_that("Init code is covered", {
     menu = function(...) 1
   )
   m_defaults_reset()
-  expect_message(llm_use())
+  expect_snapshot(llm_use(.cache = ""))
 })
 
 test_that("Stops cache", {
@@ -29,3 +29,45 @@ test_that("Stops cache", {
     llm_use("simulate_llm", "echo", .force = TRUE, .cache = "")
   )
 })
+
+test_that("Chat objects work", {
+  withr::with_envvar(
+    new = list(OPENAI_API_KEY = "test"),
+    {
+      m_defaults_reset()
+      chat <- ellmer::chat_openai()
+      expect_snapshot(
+        llm_use(chat, .cache = "")
+      )
+    }
+  )
+})
+
+test_that("Ensures empty llm_use works with Chat", {
+  withr::with_envvar(
+    new = list(OPENAI_API_KEY = "test"),
+    {
+      m_defaults_reset()
+      chat <- ellmer::chat_openai()
+      llm_use(chat, .cache = "", .silent = TRUE)
+      expect_snapshot(
+        llm_use()
+      )
+    }
+  )
+})
+
+test_that("Chat error message", {
+  withr::with_envvar(
+    new = list(OPENAI_API_KEY = "test"),
+    {
+      m_defaults_reset()
+      chat <- ellmer::chat_openai()
+      expect_error(
+        llm_use(chat, .cache = "", model = "test")
+      )
+    }
+  )
+})
+
+